# Optimizing version check to happen once per run